Update on track, Heading calculation is not working.

+++ AirTrafficMonitor/AirTrafficMonitor.Lib/Interfaces/ITrack.cs
@@ -5,7 +5,7 @@ namespace AirTrafficMonitor.Lib.Interfaces
     public interface ITrack
     {
         string Tag { get; set; }
-        int Velocity { get; set; }
+        double Velocity { get; set; }
         int Heading { get; set; }
         ICoordinate Coordinate { get; set; }
         DateTime Timestamp { get; set; }
